Player can jump, but only JumpsLeft amount of times. Camera Tracking enabled
Player is now able to jump but only according to how much we allow them to. This is done by having a trigger underneath the player and it will check for a couple of things from there. Also the default amount of jumps is now 2 for level design purposes.
Camera tracks the player but only on the vertical axis, basic lerp babY LETS GOOO